+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2023-07-05
+
+### Q1 2023 release
+
+- Updated metrics reporting so that the solution accurately sends in data for main stack deployments
+- Added ability to opt out of metrics reporting for optional function deployments (Fixity, Mediasync, Autoingest)
+
+### Contributors
+
+- @e-thoman
+
 ## [1.2.0] - 2023-06-08
 
 ### Q1 2023 release
@@ -21,7 +32,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Contributors
 
 - @e-thoman
-- @sandimciin
 
 ## [1.1.0] - 2022-07-07